sign_target_files_apks: fix recovery patch generation
While processing the target-files-package, sign_target_files_apks
maintains a temporary directory with the files needed to invoke
GetBootableImage() later in the process, needed to create the
recovery console patch. Files are passed there with the inline
write_to_temp() function.

However, in the case where the verity keys have been replaced,
the BOOT/RAMDISK/verity_key in the temporary directory is simply
omitted, it is not the updated one from ReplaceVerityPublicKey().

Similarly, the misc_info dictionary passed to GetBootableImage
still has the old value of "verity_key" in it, which means that
that boot_signer isn't being passed the correct key.

The upshot of this is the boot image data passed to
common.MakeRecoveryPatch has invalid contents and isn't being signed
with the correct key pair.

This problem is not immediately apparent as the main OTA process
will still succeed, even in an incremental update scenario, since
it doesn't care about the current state of the recovery console.
The recovery console doesn't ever get updated however since the hashes
don't match up, and errors appear in the logcat at first boot after
the update due to this issue.

Change-Id: I1643210d0e8a10dae9246f7ef9e5218d709454c3
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/337123
Signed-off-by: Cody Xie <xkd@rock-chips.com>